// src/Rest.type.ts
/**
 * @author gye hyun james kim <pnuskgh@gmail.com>
 * @copyright 2017~2025, OBCon Inc.
 * @license GNU GENERAL PUBLIC LICENSE v3.0 (https://github.com/jopenbusiness/StockHub?tab=GPL-3.0-1-ov-file)
 */

export interface RETURN_INFO {
    code: number;                                           //--- 응답 코드
    message: string;                                        //--- 응답 메시지
    // eslint-disable-next-line @typescript-eslint/no-explicit-any
    data?: any;                                             //--- 응답 데이터
    // eslint-disable-next-line @typescript-eslint/no-explicit-any
    error?: any;                                            //--- 오류 데이터
};